docs: Update remote-notification-support.md
Added warning about mutable-content requirement for NSE to work and also about content-available so the notifications works when app is in foreground.

Also edited section for Edit Payload to include second option on how to edit the payload. Second option is kind of better imho as BE do not need to care too much about FE implementation and it will be easier to switch to other libs - only FE needs to update the code and BE can stay same